refactor: LightingPanel setup into its own class
